Move config to own package
